move init useEffect() into useReducerHook and add reducer
