+++ src/contexts/contacts.context.js
@@ -1,6 +1,8 @@
 import { createContext, useContext, useEffect, useState } from 'react';
 import Loading from '../components/Loading';
 
+import contactReducer from '../reducers/contact.reducer';
+import useContactReducer from '../hooks/useContactReducer';
 import fetchDatabase from '../services/api/fetchDatabase';
 import { useLayout } from './layout.context';
 import { useUser } from './user.context';
@@ -37,27 +39,11 @@ function ContactsProvider({ children }) {
 	const { contacts, uuid } = useUser();
 	const { dataListTab } = useLayout();
 	const [foundContactsData, setFoundContactsData] = useState([]);
-	const [contactsData, setContactsData] = useState('');
-	const [isFetching, setIsFetching] = useState(true);
-
-	useEffect(() => {
-		setIsFetching(true);
-		const activeContacts = Object.keys(contacts).filter(key => contacts[key] === true);
-		const fetchedData = activeContacts.map(contact => {
-			return fetchDatabase(`/users/${contact}`)
-				.then(data => {
-					const contactObj = {
-						displayName: data.displayName,
-						uuid: data.uuid,
-						photoURL: data.photoURL,
-					};
-					return contactObj;
-				})
-				.catch(error => console.log(error));
-		});
-		Promise.all(fetchedData).then(data => setContactsData(data));
-		setIsFetching(false);
-	}, [contacts]);
+	const [contactsData, dispatch, isFetching] = useContactReducer(
+		contactReducer,
+		contacts,
+		[]
+	);
 
 	useEffect(() => {
 		if (dataListTab.contacts === 'findContacts') {
@@ -79,9 +65,11 @@ function ContactsProvider({ children }) {
 
 	return (
 		<ContactsContext.Provider value={contactsData}>
-			<FindContactsContext.Provider value={foundContactsData}>
-				{!isFetching ? children : <Loading />}
-			</FindContactsContext.Provider>
+			<ContactsDispatch.Provider value={dispatch}>
+				<FindContactsContext.Provider value={foundContactsData}>
+					{!isFetching ? children : <Loading />}
+				</FindContactsContext.Provider>
+			</ContactsDispatch.Provider>
 		</ContactsContext.Provider>
 	);
 }
